+++ lib/src/chart/index.ts
@@ -1,2 +1,2 @@
 export { ChartWrapper as Chart } from "./ChartWrapper";
-export type { ChartApiRef, ChartProps, ChartCustomOptions } from "./types";
+export type { ChartApiRef, ChartProps } from "./types";