+++ lib/src/chart/types.ts
@@ -7,17 +7,14 @@ import {
 } from "lightweight-charts";
 import type { JSX, ReactNode } from "react";
 
-export type ChartCustomOptions = {
+export type ChartProps = {
+  children?: ReactNode;
+  containerProps?: JSX.IntrinsicElements["div"];
   onClick?: MouseEventHandler<Time>;
   onCrosshairMove?: MouseEventHandler<Time>;
   onInit?: (chart: IChartApi) => void;
   options?: DeepPartial<ChartNativeOptions>;
 };
-
-export type ChartProps = {
-  children?: ReactNode;
-  containerProps?: JSX.IntrinsicElements["div"];
-} & ChartCustomOptions;
 
 export type ChartApiRef = {
   _chart: IChartApi | null;
@@ -30,3 +27,11 @@ export interface IChartContext {
   chartApiRef: ChartApiRef | null;
   isReady: boolean;
 }
+
+export type ChartComponentProps = {
+  container: HTMLElement;
+} & Omit<ChartProps, "containerProps">;
+
+export type UseChartOptions = {
+  container: HTMLElement;
+} & Omit<ChartProps, "children" | "containerProps">;