+++ lib/CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Feat
+- fix chart component incorrect typngs (#64)
 
-## [0.3.0] - 2025-04-08
+## [0.3.0] - 2025-04-21
 ### Feat
 - add `containerProps` and `options` to Chart props
 - add `isPane` series prop